+++ src/SmartContorl/MultipleTargetGeneticAlgorithmUI.cpp
@@ -258,18 +258,18 @@ void MultipleTargetGeneticAlgorithmUI::on_pushButtonTargetChart_clicked()
 		listValues.push_back(valuePtr);
 	}
 
-	std::vector<std::map<QString, std::vector<double>>> parValues;
+	std::vector<std::map<QString, std::vector<float>>> parValues;
 	for (int i = 0; i < valueCount; i++)
 	{
-		std::map<QString, std::vector<double>> parValue;
+		std::map<QString, std::vector<float>> parValue;
 		for (auto iter = histroy.begin()->variates.begin(); iter != histroy.begin()->variates.end(); iter++)
 		{
-			parValue[iter->name] = std::vector<double>();
+			parValue[iter->name] = std::vector<float>();
 		}
 
 		for (int t = 0; t < targetCount; t++)
 		{
-			parValue[QString("F%1").arg(t)] = std::vector<double>();
+			parValue[QString("F%1").arg(t)] = std::vector<float>();
 		}
 
 		parValues.push_back(parValue);
@@ -455,17 +455,17 @@ void MultipleTargetGeneticAlgorithmUI::on_pushButtonVariateMax_clicked()
 		listValues.push_back(valuePtr);
 	}
 
-	std::vector<std::map<QString, std::vector<double>>> parValues;
+	std::vector<std::map<QString, std::vector<float>>> parValues;
 	for (int i = 0; i < valueCount; i++)
 	{
-		auto parV = std::map<QString, std::vector<double>>();
+		auto parV = std::map<QString, std::vector<float>>();
 		for (int j = 0; j < variates.size(); j++)
 		{
 			if (j == selectIndex)
 				continue;
-			parV[variates[j].name] = std::vector<double>();
+			parV[variates[j].name] = std::vector<float>();
 		}
-		parV["F"] = std::vector<double>();
+		parV["F"] = std::vector<float>();
 		parValues.push_back(parV);
 	}
 	int temp = 1;
